requirements.txt
asgiref==3.6.0
Django==4.1.6
django-browser-reload==1.6.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==6.0.0
Markdown==3.4.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.2
whitenoise==6.2.0
zipp==3.12.1
